+++ Docker_Python/build.sh
@@ -1,10 +1,9 @@
-#docker-compose -f docker-compose.yml build
+docker-compose -f docker-compose.yml build
 #docker-compose -f docker-compose.yml build light
 #docker-compose -f docker-compose.yml build default
 #docker-compose -f docker-compose.yml build full
 #docker-compose -f docker-compose.yml build scvi
 #docker-compose -f docker-compose.yml build rapidsc
-#exit
 
 #reponame=shortcake
 #docker save -o $reponame.tar rnakato/$reponame
@@ -12,7 +11,7 @@
 
 for name in shortcake_light shortcake shortcake_full shortcake_scvi shortcake_rapidsc
 do
-    tag=3.2.0
+    tag=3.3.0
     docker tag rnakato/$name:$tag rnakato/$name:latest
 #    docker save -o $name-$tag.tar rnakato/$name:$tag
 #    singularity build -F /work3/SingularityImages/$name.$tag.sif docker-archive://$name-$tag.tar
@@ -22,3 +21,4 @@ do
          docker push rnakato/$name:$tag
     done
 done
+